+++ server/src/lti/views.py
@@ -4,12 +4,14 @@ from django.conf import settings
 
 from lti.services.LTIService import validate_lti_request, request_to_course, request_to_user, create_course_user
 from users.services.TokenService import generate_token
+from ripple.util.util import get_root_path
 
 
 def index(request):
     lti_params = request.POST
     lti_validation = validate_lti_request(
         request.build_absolute_uri(), request.method, lti_params)
+    root_path = get_root_path(request)
 
     if lti_validation.get("error", None) is not None:
         user_response = {
@@ -17,7 +19,7 @@ def index(request):
         }
         return JsonResponse(user_response)
     else:
-        user = request_to_user(lti_params)
+        user = request_to_user(lti_params, root_path)
         course = request_to_course(lti_params, user)
         course_user = create_course_user(course, user, lti_params)
         # Get token
