+++ server/src/ripple/settings.py
@@ -177,6 +177,7 @@ ACHIEVEMENT_CLASSES = ['users.achievements.userAchievements']
 
 
 RUNTIME_CONFIGURATION = {
+    "default_user_image": "default_images/default_user_image.png",
     "min_competency_threshold": 3,
     "min_competency_value": 0.9,
     "report_reason_list": ["custom", "Inappropriate Content", "Incorrect Answer",  "Incorrect Tags"],
